frontend/requirements.txt
fastapi==0.115.5
uvicorn
httpx
python-fasthtml==0.10.0
shad4fast==1.3.1
python-multipart
sqlite-minutils